+++ pipeline/outputmodules/validate_test.go
@@ -2,12 +2,25 @@ package outputmodules
 
 import (
 	"fmt"
+	"testing"
+
 	pbsubstreams "github.com/streamingfast/substreams/pb/sf/substreams/v1"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func Test_ValidateRequest(t *testing.T) {
+	testOutputMap := withOutputModule("output_mod", "map", false)
+	testOutputStore := withOutputModule("output_mod", "store", false)
+	testOutputMapLegacy := withOutputModule("output_mod", "map", true)
+	testOutputStoreLegacy := withOutputModule("output_mod", "store", true)
+
+	testBlockType := "sf.substreams.v1.test.Block"
+
+	stepNew := pbsubstreams.ForkStep_STEP_NEW
+	stepUndo := pbsubstreams.ForkStep_STEP_UNDO
+	stepIrreversible := pbsubstreams.ForkStep_STEP_IRREVERSIBLE
+	stepUnknown := pbsubstreams.ForkStep_STEP_UNKNOWN
+
 	tests := []struct {
 		name       string
 		request    *pbsubstreams.Request
@@ -15,25 +28,40 @@ func Test_ValidateRequest(t *testing.T) {
 		blockType  string
 		expect     error
 	}{
-		{"negative start block num", req(-1), false, "sf.substreams.v1.test.Block", fmt.Errorf("negative start block -1 is not accepted")},
-		{"no modules found in request", &pbsubstreams.Request{StartBlockNum: 1}, false, "sf.substreams.v1.test.Block", fmt.Errorf("no modules found in request")},
-		{"multiple output modules is not accepted", req(1, withOutputModules([][]string{{"output_mod_1", "store"}, {"output_mod_1", "kind"}}, true)), false, "sf.substreams.v1.test.Block", fmt.Errorf("multiple output modules is not accepted")},
-		{"single legacy map output module is accepted for none sub-request", req(1, withOutputModule("output_mod", "map", true)), false, "sf.substreams.v1.test.Block", nil},
-		{"single legacy store output module is not accepted for none sub-request", req(1, withOutputModule("output_mod", "store", true)), false, "sf.substreams.v1.test.Block", fmt.Errorf("multiple output modules is not accepted")},
-		{"single legacy map output module is accepted for sub-request", req(1, withOutputModule("output_mod", "map", true)), true, "sf.substreams.v1.test.Block", nil},
-		{"single legacy store output module is accepted for sub-request", req(1, withOutputModule("output_mod", "store", true)), true, "sf.substreams.v1.test.Block", nil},
-		{"single map output module is accepted for none sub-request", req(1, withOutputModule("output_mod", "map", false)), false, "sf.substreams.v1.test.Block", nil},
-		{"single store output module is not accepted for none sub-request", req(1, withOutputModule("output_mod", "store", false)), false, "sf.substreams.v1.test.Block", fmt.Errorf("multiple output modules is not accepted")},
-		{"single map output module is accepted for none sub-request", req(1, withOutputModule("output_mod", "map", false)), true, "sf.substreams.v1.test.Block", nil},
-		{"single store output module is  accepted for none sub-request", req(1, withOutputModule("output_mod", "map", false)), true, "sf.substreams.v1.test.Block", nil},
-		{name: "debug initial snapshots not accepted in production mode", request: req(1, withDebugInitialSnapshotForModules([]string{"foo"}), withProductionMode()), expect: fmt.Errorf("debug initial snapshots not accepted in production mode")},
+		{"negative start block num", req(-1, testOutputMap), false, testBlockType, nil},
+		{"no modules found in request", &pbsubstreams.Request{StartBlockNum: 1}, false, testBlockType, fmt.Errorf("modules validation failed: no modules found in request")},
+		{"multiple output modules is not accepted", req(1, withOutputModules([][]string{{"output_mod_1", "store"}, {"output_mod_1", "kind"}}, true)), false, testBlockType, fmt.Errorf("validate request: output module: multiple output modules is not accepted")},
+		{"single legacy map output module is accepted for none sub-request", req(1, testOutputMapLegacy), false, testBlockType, nil},
+		{"single legacy store output module is not accepted for none sub-request", req(1, testOutputStoreLegacy), false, testBlockType, fmt.Errorf("validate request: output module must be of kind 'map'")},
+		{"single legacy map output module is accepted for sub-request", req(1, testOutputMapLegacy), true, testBlockType, nil},
+		{"single legacy store output module is accepted for sub-request", req(1, testOutputStoreLegacy), true, testBlockType, nil},
+		{"single map output module is accepted for none sub-request", req(1, testOutputMap), false, testBlockType, nil},
+		{"single store output module is not accepted for none sub-request", req(1, testOutputStore), false, testBlockType, fmt.Errorf("validate request: output module must be of kind 'map'")},
+		{"single map output module is accepted for none sub-request", req(1, testOutputMap), true, testBlockType, nil},
+		{"single store output module is  accepted for none sub-request", req(1, testOutputMap), true, testBlockType, nil},
+		{"debug initial snapshots not accepted in production mode", req(1, withDebugInitialSnapshotForModules([]string{"foo"}), withProductionMode()), false, "", fmt.Errorf("debug initial store snapshot feature is not supported in production mode")},
+
+		{"step empty rejected", req(1, testOutputMap, withSteps()), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": cannot be empty`)},
+
+		{"step undo rejected", req(1, testOutputMap, withSteps(stepUndo)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": step "STEP_UNDO" cannot be specified alone`)},
+		{"step unknown rejected", req(1, testOutputMap, withSteps(stepUnknown)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": step "STEP_UNKNOWN" cannot be specified alone`)},
+
+		{"step only new/undo accepted", req(1, testOutputMap, withSteps(stepNew, stepUndo)), false, testBlockType, nil},
+		{"step only undo/new accepted", req(1, testOutputMap, withSteps(stepUndo, stepNew)), false, testBlockType, nil},
+		{"step only irreversible accepted", req(1, testOutputMap, withSteps(stepIrreversible)), false, testBlockType, nil},
+
+		{"step new/unknown rejected", req(1, testOutputMap, withSteps(stepNew, stepUnknown)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": step "STEP_NEW" and step "STEP_UNKNOWN" cannot be provided together accepting "STEP_NEW" and "STEP_UNDO" only`)},
+		{"step new/irreversible rejected", req(1, testOutputMap, withSteps(stepNew, stepIrreversible)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": step "STEP_NEW" and step "STEP_IRREVERSIBLE" cannot be provided together accepting "STEP_NEW" and "STEP_UNDO" only`)},
+		{"step irreversible/undo rejected", req(1, testOutputMap, withSteps(stepIrreversible, stepUndo)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": step "STEP_IRREVERSIBLE" and step "STEP_UNDO" cannot be provided together accepting "STEP_NEW" and "STEP_UNDO" only`)},
+
+		{"step new/undo/irreversible rejected", req(1, testOutputMap, withSteps(stepNew, stepUndo, stepIrreversible)), false, testBlockType, fmt.Errorf(`validate request: invalid "fork_steps": accepting only 1 or 2 steps but there was 3 steps provided`)},
 	}
 
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			err := ValidateRequest(test.request, test.blockType, test.subrequest)
 			if test.expect != nil {
-				require.Error(t, err)
+				require.EqualError(t, err, test.expect.Error())
 			} else {
 				require.NoError(t, err)
 			}
@@ -59,6 +87,13 @@ func withOutputModule(outputModule, kind string, legacy bool) reqOption {
 	}
 }
 
+func withSteps(steps ...pbsubstreams.ForkStep) reqOption {
+	return func(req *pbsubstreams.Request) *pbsubstreams.Request {
+		req.ForkSteps = steps
+		return req
+	}
+}
+
 func withProductionMode() reqOption {
 	return func(req *pbsubstreams.Request) *pbsubstreams.Request {
 		req.ProductionMode = true
@@ -77,6 +112,7 @@ func req(startBlockNum int64, opts ...reqOption) *pbsubstreams.Request {
 	r := &pbsubstreams.Request{
 		StartBlockNum: startBlockNum,
 		Modules:       &pbsubstreams.Modules{},
+		ForkSteps:     []pbsubstreams.ForkStep{pbsubstreams.ForkStep_STEP_NEW, pbsubstreams.ForkStep_STEP_UNDO},
 	}
 	for _, opt := range opts {
 		r = opt(r)