+++ test/runnable_test.go
@@ -76,6 +76,7 @@ func (f *testRun) Run(t *testing.T) error {
 		StartBlockNum:  f.StartBlock,
 		StopBlockNum:   f.ExclusiveEndBlock,
 		StartCursor:    opaqueCursor,
+		ForkSteps:      []pbsubstreams.ForkStep{pbsubstreams.ForkStep_STEP_NEW, pbsubstreams.ForkStep_STEP_UNDO},
 		Modules:        pkg.Modules,
 		OutputModule:   f.ModuleName,
 		ProductionMode: f.ProductionMode,
@@ -257,7 +258,7 @@ type TestRunner struct {
 	generator TestBlockGenerator
 }
 
-func (r *TestRunner) StreamFactory(_ context.Context, h bstream.Handler, startBlockNum int64, stopBlockNum uint64, _ string, _ bool) (service.Streamable, error) {
+func (r *TestRunner) StreamFactory(_ context.Context, h bstream.Handler, startBlockNum int64, stopBlockNum uint64, _ string, _ bool, _ bool) (service.Streamable, error) {
 	r.pipe = h.(*pipeline.Pipeline)
 	r.Shutter = shutter.New()
 	r.generator = r.blockGeneratorFactory(uint64(startBlockNum), stopBlockNum)